backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dill==0.4.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
isort==6.0.1
lxml==6.0.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.7
python-cas==1.6.0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
